[IMP] sale_quotation_builder: Split files according to guidelines
