+++ addons/sale_quotation_builder/models/__init__.py
@@ -4,4 +4,6 @@
 from . import product_template
 from . import res_company
 from . import sale_order
+from . import sale_order_line
+from . import sale_order_option
 from . import sale_order_template
